[step2] fix: 로또 당첨 개수가 3개 이하인 경우에 대한 전체 당첨 개수 로직 수정
